+++ src/youtube_automation/agents/_preflight.py
@@ -1,0 +1,130 @@
+"""アップロード前メタデータ品質チェック（fail-loud preflight）。
+
+``YouTubeAutoUploader`` から分離した mixin。挙動は分割前と同一。
+``self._extract_md_section`` は ``DescriptionsMdMixin`` が提供する。
+"""
+
+from __future__ import annotations
+
+import json
+import logging
+import re
+from pathlib import Path
+
+from youtube_automation.utils.collection_paths import CollectionPaths
+from youtube_automation.utils.config import load_config
+from youtube_automation.utils.preflight_checks import (
+    check_chapter_count,
+    check_chapter_variation_suffix,
+    check_duration,
+    check_low_cpm_localization_languages,
+    check_required_localization_languages,
+    check_tags_count,
+    check_tags_yt_chars,
+    extract_descriptions_md_tags,
+)
+from youtube_automation.utils.probe import probe_duration
+
+logger = logging.getLogger(__name__)
+
+
+class PreflightMixin:
+    """アップロード前メタデータ検証を提供する mixin。"""
+
+    def _preflight_check(self, collection_dir: Path) -> None:
+        """アップロード前メタデータ品質チェック (fail-loud)。
+
+        過去事例の再発防止:
+        1. descriptions.md が存在すること（Track 01 仮名フォールバックを防ぐ）
+        2. workflow-state.json.scene_phrases に EN + 全 supported_languages が
+           揃っていること（多言語タイトルが EN ベタコピーになる事故を防ぐ）
+        3. タイムスタンプ件数が `audio.chapter_max` 以内かつ chapter 名に
+           パターン展開接尾辞（v1〜v6 / ロマン数字 I〜VIII）を含まないこと
+           （個別トラック = 1 chapter の per-track 命名はデフォルトで許容）
+        4. タイトルが 100 codepoint 以内（YouTube 制限）
+        5. タグ件数が `tags.min_count` を満たすこと（戦略書違反防止）
+        6. タグの quotation 込み文字数が YouTube の 500 制限内
+        7. master 動画尺が `audio.target_duration_min/max` 範囲内
+        8. supported_languages が高 CPM 必須言語 ja/en/de を含むこと
+        """
+        paths = CollectionPaths(collection_dir)
+        desc_path = paths.descriptions_md_path
+        if not desc_path.exists():
+            raise RuntimeError(f"❌ {desc_path} が存在しません。/video-description を実行してください。")
+
+        text = desc_path.read_text(encoding="utf-8")
+        title = (self._extract_md_section(text, "タイトル案") or "").strip()
+        description = (self._extract_md_section(text, "Complete Collection 概要欄") or "").strip()
+
+        if not title or not description:
+            raise RuntimeError(f"❌ {desc_path}: タイトル案 / Complete Collection 概要欄 が空")
+
+        if len(title) > 100:
+            raise RuntimeError(f"❌ タイトルが {len(title)} codepoint。YouTube 制限 100 を超過。\n  {title}")
+
+        config = load_config()
+        msg = check_required_localization_languages(config.localizations.supported_languages)
+        if msg:
+            raise RuntimeError(f"❌ {msg}。config/localizations.json を見直してください。")
+        msg = check_low_cpm_localization_languages(config.localizations.supported_languages)
+        if msg:
+            logger.warning(f"⚠️  {msg}。意図的な例外でなければ config/localizations.json を見直してください。")
+
+        # タイムスタンプ粒度検証
+        ts_lines = [line for line in description.split("\n") if re.match(r"^\d{1,2}:\d{2}", line.strip())]
+        if len(ts_lines) < 3:
+            raise RuntimeError(f"❌ タイムスタンプ {len(ts_lines)} 個 (最低 3 必要)")
+        msg = check_chapter_count(len(ts_lines), config.audio.chapter_max)
+        if msg:
+            raise RuntimeError(f"❌ {msg}。config.audio.chapter_max を見直してください。")
+        msg = check_chapter_variation_suffix(ts_lines)
+        if msg:
+            raise RuntimeError(f"❌ {msg}: 1 パターン = 1 chapter で再生成してください。")
+
+        # scene_phrases 完全性検証
+        ws_path = paths.workflow_state_path
+        state = json.loads(ws_path.read_text(encoding="utf-8")) if ws_path.exists() else {}
+        scene_phrases = state.get("scene_phrases") or {}
+
+        required_langs = ["en"] + list(config.localizations.supported_languages)
+        missing = [lang for lang in required_langs if not scene_phrases.get(lang)]
+        if missing:
+            raise RuntimeError(
+                f"❌ workflow-state.json.scene_phrases に翻訳が不足: {missing}\n"
+                f"→ /video-description で多言語翻訳を含めて再生成してください。\n"
+                f"→ 既存例: collections/live/20260322-rjn-city-collection/workflow-state.json"
+            )
+
+        # タグ件数 / quotation 文字数チェック
+        # descriptions.md の「タグ（YouTube タグ欄）」が _upload_complete_collection で
+        # for_collection() を上書きするため、本番と同じソースを検証する。
+        prebuilt_tags = extract_descriptions_md_tags(desc_path)
+        tags = prebuilt_tags if prebuilt_tags is not None else config.content.tags.for_collection(collection_dir.name)
+        issues: list[str] = []
+        for msg in (
+            check_tags_count(tags, config.content.tags.min_count),
+            check_tags_yt_chars(tags),
+        ):
+            if msg:
+                issues.append(msg)
+
+        # 動画尺チェック（target_duration が設定済みかつ master mp4 が存在する場合のみ）
+        if config.audio.target_duration_min is not None or config.audio.target_duration_max is not None:
+            master_video = paths.find_master_video()
+            if master_video:
+                dur = probe_duration(master_video)
+                if dur is None:
+                    issues.append(f"duration probe failed for {master_video.name}")
+                else:
+                    msg = check_duration(
+                        dur,
+                        config.audio.target_duration_min,
+                        config.audio.target_duration_max,
+                    )
+                    if msg:
+                        issues.append(msg)
+
+        if issues:
+            raise RuntimeError("❌ preflight failed:\n  - " + "\n  - ".join(issues))
+
+        logger.info(f"✅ preflight OK — title={len(title)}c, chapters={len(ts_lines)}, langs={len(scene_phrases)}")
